+++ bevy_rtc/src/client/mod.rs
@@ -8,5 +8,5 @@ mod systems;
 pub use events::{ConnectionRequest, RtcClientEvent};
 pub use plugin::RtcClientPlugin;
 pub use router::AddProtocolExt;
-pub use state::{RtcClientStatus, RtcState};
-pub use system_params::{NetworkReader, NetworkWriter};
+pub use state::{RtcClientState, RtcClientStatus};
+pub use system_params::RtcClient;